+++ src/__generated__/QueryRenderersBidFlowQuery.graphql.ts
@@ -63,6 +63,8 @@ fragment SelectMaxBid_sale_artwork on SaleArtwork {
 fragment ConfirmBid_sale_artwork on SaleArtwork {
   sale {
     id
+    live_start_at
+    end_at
     __id
   }
   artwork {
@@ -80,6 +82,8 @@ fragment ConfirmBid_sale_artwork on SaleArtwork {
 fragment ConfirmFirstTimeBid_sale_artwork on SaleArtwork {
   sale {
     id
+    live_start_at
+    end_at
     __id
   }
   artwork {
@@ -284,7 +288,6 @@ return {
                 "plural": false,
                 "selections": [
                   v6,
-                  v3,
                   {
                     "kind": "ScalarField",
                     "alias": null,
@@ -298,7 +301,8 @@ return {
                     "name": "end_at",
                     "args": null,
                     "storageKey": null
-                  }
+                  },
+                  v3
                 ]
               },
               {