+++ src/lib/relay/QueryRenderers.tsx
@@ -59,11 +59,13 @@ export const ArtistRenderer: React.SFC<ArtistRendererProps> = ({ render, artistI
 }
 
 interface BidFlowRendererProps extends RendererProps {
-  artworkID: string
+  artworkID?: string
   saleID: string
+  intent: "bid" | "register"
 }
 
-export const BidFlowRenderer: React.SFC<BidFlowRendererProps> = ({ render, artworkID, saleID }) => {
+export const BidFlowRenderer: React.SFC<BidFlowRendererProps> = ({ render, artworkID, saleID, intent }) => {
+  // TODO: artworkID can be nil, so omit that part of the query if it is.
   return (
     <QueryRenderer
       environment={environment}
@@ -93,6 +95,7 @@ export const BidFlowRenderer: React.SFC<BidFlowRendererProps> = ({ render, artwo
             props: {
               sale_artwork: props.artwork.sale_artwork,
               me: props.me,
+              intent,
             },
             error,
           })